Ajustes baseactivity
